requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.3.0
certifi==2025.1.31
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
crispy-bootstrap4==2024.10
crispy-bootstrap5==2025.6
decorator==5.1.1
distlib==0.3.9
Django==4.2.17
django-cors-headers==4.7.0
django-crispy-forms==2.3
django-dbbackup==4.3.0
django-filter==24.3
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.10
filelock==3.18.0
frozenlist==1.6.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
gps==3.19
httplib2==0.22.0
idna==3.10
inflection==0.5.1
mailjet-rest==1.4.0
msgpack==1.1.0
multidict==6.4.3
mysqlclient==2.2.7
oauth2client==4.1.3
packaging==25.0
paynow==1.0.8
pillow==11.0.0
platformdirs==4.3.8
propcache==0.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
PyJWT==2.10.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
ratelim==0.1.6
redis==5.2.1
requests==2.32.3
rest-framework-simplejwt==0.0.2
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
twilio==9.6.0
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
virtualenv==20.31.2
yarl==1.20.0
